Installation

Installation

To install, run

] add SimpleDifferentialOperators

Note that this requires Julia 1.0 or later.

For the complete derivations of the objects we return, see the PDF.

Formulas

Under $P$-length of grids on $v$ with end points $\underline{z} < \overline{z}$ with reflecting barrier conditions of

\[\begin{align} \underline{\xi} v(\underline{z}) + \nabla v(\underline{z} ) &= 0\\ \overline{\xi} v(\overline{z}) + \nabla v(\overline{z}) &= 0 \end{align}\]

We use the following discretization schemes:

\[L_1^{-} \equiv \frac{1}{\Delta}\begin{pmatrix} 1 - (1 + \underline{\xi} \Delta) &0&0&\dots&0&0&0\\ -1&1&0&\dots&0&0&0\\ \vdots&\vdots&\vdots&\ddots&\vdots&\vdots&\vdots\\ 0&0&0&\dots&-1&1&0\\ 0&0&0&\cdots&0&-1&1 \end{pmatrix}_{P\times P}\]

\[L_1^{+} \equiv \frac{1}{\Delta}\begin{pmatrix} -1&1&0&\dots&0&0&0\\ 0&-1&1&\dots&0&0&0\\ \vdots&\vdots&\vdots&\ddots&\vdots&\vdots&\vdots\\ 0&0&0&\dots&0&-1&1\\ 0&0&0&\cdots&0&0&-1+(1-\overline{\xi} \Delta) \end{pmatrix}_{P\times P}\label{eq:L-1-plus-regular} \\\]

\[L_2 \equiv \frac{1}{\Delta^2}\begin{pmatrix} -2 + (1 + \underline{\xi} \Delta) &1&0&\dots&0&0&0\\ 1&-2&1&\dots&0&0&0\\ \vdots&\vdots&\vdots&\ddots&\vdots&\vdots&\vdots\\ 0&0&0&\dots&1&-2&1\\ 0&0&0&\cdots&0&1&-2 + (1- \overline{\xi} \Delta) \end{pmatrix}_{P\times P}\label{eq:L-2-regular}\]

which represent the backward first order, foward first order, and central second order differential operators respectively.

Derivation, including formula for irregular grids, can be found here.

Usage

This package provides discretized differential operators of first order and second order under reflecting boundary conditions.

SimpleDifferentialOperators.diffusionoperators — Method.

diffusionoperators(x, BC1::BoundaryCondition, BC2::BoundaryCondition)

Return the diffusion operators (L_1_minus, L_1_plus, L_2) w.r.t the supplied grid and BCs.

x is a grid (either an AbstractRange, in which we use specialized uniform grid code, or an AbstractArray). The first BC binds at the lower end of the grid (i.e., x[1]), and the latter at the high end. The BCs are either a Reflecting(), or "Dirichlet" boundary condition v'(x) = 0, or Mixed(x::T) where {T <: Real}, corresponding to "Robin" boundary conditions.

source