API

API

SimpleDifferentialOperators.diffusionoperators

SimpleDifferentialOperators.diffusionoperators

SimpleDifferentialOperators.diffusionoperators — Method.

diffusionoperators(x, BC1::BoundaryCondition, BC2::BoundaryCondition)

Returns a tuple of diffusion operators and extended grid (L_1_minus, L_1_plus, L_2, x_bar) with specified boundary conditions.

Given a grid x of length M, return diffusion operators for negative drift, positive drift, and central differences. BC1 is applied to the lower bound, and BC2 to the upper. x_bar is a (M+2) array that represents the extended grid whose first and last elements represent the ghost nodes just before x[1] and x[end].

Examples

julia> x = 1:3
1:3

julia> L_1_minus, L_1_plus, L_2, x_bar = diffusionoperators(x, Reflecting(), Reflecting());

julia> Array(L_1_minus)
3×3 Array{Float64,2}:
  0.0   0.0  0.0
 -1.0   1.0  0.0
  0.0  -1.0  1.0

julia> Array(L_1_plus)
3×3 Array{Float64,2}:
 -1.0   1.0  0.0
  0.0  -1.0  1.0
  0.0   0.0  0.0

julia> Array(L_2)
3×3 Array{Float64,2}:
 -1.0   1.0   0.0
  1.0  -2.0   1.0
  0.0   1.0  -1.0

julia> x_bar
5-element Array{Int64,1}:
 0
 1
 2
 3
 4

source

SimpleDifferentialOperators.diffusionoperators — Method.

diffusionoperators(x)

Returns a tuple of diffusion operators and extended grid (L_1_minus, L_1_plus, L_2, x_bar) without applying any boundary conditions.

Given a grid x of length M, return diffusion operators for negative drift, positive drift, and central differences. No boundary conditions are applied. x_bar is a (M+2) array that represents the extended grid whose first and last elements represent the ghost nodes just before x[1] and x[end].

Examples

julia> x = 1:3
1:3

julia> L_1_minus, L_1_plus, L_2, x_bar = diffusionoperators(x);

julia> @show Array(L_1_minus)
Array(L_1_minus) = [-1.0 1.0 0.0 0.0 0.0; 0.0 -1.0 1.0 0.0 0.0; 0.0 0.0 -1.0 1.0 0.0]
3×5 Array{Float64,2}:
 -1.0   1.0   0.0  0.0  0.0
  0.0  -1.0   1.0  0.0  0.0
  0.0   0.0  -1.0  1.0  0.0

julia> @show Array(L_1_plus)
Array(L_1_plus) = [0.0 -1.0 1.0 0.0 0.0; 0.0 0.0 -1.0 1.0 0.0; 0.0 0.0 0.0 -1.0 1.0]
3×5 Array{Float64,2}:
 0.0  -1.0   1.0   0.0  0.0
 0.0   0.0  -1.0   1.0  0.0
 0.0   0.0   0.0  -1.0  1.0

julia> @show Array(L_2)
Array(L_2) = [1.0 -2.0 1.0 0.0 0.0; 0.0 1.0 -2.0 1.0 0.0; 0.0 0.0 1.0 -2.0 1.0]
3×5 Array{Float64,2}:
 1.0  -2.0   1.0   0.0  0.0
 0.0   1.0  -2.0   1.0  0.0
 0.0   0.0   1.0  -2.0  1.0

julia> @show x_bar
x_bar = [0, 1, 2, 3, 4]
5-element Array{Int64,1}:
 0
 1
 2
 3
 4

source