API

API

SimpleDifferentialOperators.DifferentialOperator

SimpleDifferentialOperators.DifferentialOperator

SimpleDifferentialOperators.ExtensionDifferentialOperator

SimpleDifferentialOperators.L₁₊

SimpleDifferentialOperators.L₁₊bc

SimpleDifferentialOperators.L₁₋

SimpleDifferentialOperators.L₁₋bc

SimpleDifferentialOperators.L₂

SimpleDifferentialOperators.L₂bc

SimpleDifferentialOperators.extrapolatetoboundary

SimpleDifferentialOperators.extrapolatetoboundary

SimpleDifferentialOperators.interiornodes

SimpleDifferentialOperators.interiornodes

SimpleDifferentialOperators.DifferentialOperator — Method.

DifferentialOperator(x̄, bc::Tuple{Mixed, Mixed}, method::DifferenceMethod)

Returns a discretized differential operator of length(x̄) by length(x̄) matrix under mixed boundary conditions from bc using finite difference method specified by method.

Examples

julia> x̄ = 0:5
0:5

julia> DifferentialOperator(x̄, (Mixed(ξ = 1.0), Mixed(ξ = 1.0)), BackwardFirstDifference())
4×4 SparseArrays.SparseMatrixCSC{Float64,Int64} with 7 stored entries:
  [1, 1]  =  Inf
  [2, 1]  =  -1.0
  [2, 2]  =  1.0
  [3, 2]  =  -1.0
  [3, 3]  =  1.0
  [4, 3]  =  -1.0
  [4, 4]  =  1.0

julia> DifferentialOperator(x̄, (Mixed(ξ = 1.0), Mixed(ξ = 1.0)), ForwardFirstDifference())
4×4 SparseArrays.SparseMatrixCSC{Float64,Int64} with 7 stored entries:
  [1, 1]  =  -1.0
  [1, 2]  =  1.0
  [2, 2]  =  -1.0
  [2, 3]  =  1.0
  [3, 3]  =  -1.0
  [3, 4]  =  1.0
  [4, 4]  =  -0.5

julia> DifferentialOperator(x̄, (Mixed(ξ = 1.0), Mixed(ξ = 1.0)), CentralSecondDifference())
4×4 SparseArrays.SparseMatrixCSC{Float64,Int64} with 10 stored entries:
  [1, 1]  =  -Inf
  [2, 1]  =  1.0
  [1, 2]  =  1.0
  [2, 2]  =  -2.0
  [3, 2]  =  1.0
  [2, 3]  =  1.0
  [3, 3]  =  -2.0
  [4, 3]  =  1.0
  [3, 4]  =  1.0
  [4, 4]  =  -1.5

source

SimpleDifferentialOperators.DifferentialOperator — Method.

DifferentialOperator(x̄, bc::Tuple{Reflecting, Reflecting}, method::DifferenceMethod)

Returns a discretized differential operator of length(x̄) by length(x̄) matrix under reflecting boundary conditions from bc using finite difference method specified by method.

Examples

julia> x̄ = 0:5
0:5

julia> DifferentialOperator(x̄, (Reflecting(), Reflecting()), BackwardFirstDifference())
4×4 SparseArrays.SparseMatrixCSC{Float64,Int64} with 7 stored entries:
  [1, 1]  =  0.0
  [2, 1]  =  -1.0
  [2, 2]  =  1.0
  [3, 2]  =  -1.0
  [3, 3]  =  1.0
  [4, 3]  =  -1.0
  [4, 4]  =  1.0

julia> DifferentialOperator(x̄, (Reflecting(), Reflecting()), ForwardFirstDifference())
4×4 SparseArrays.SparseMatrixCSC{Float64,Int64} with 7 stored entries:
  [1, 1]  =  -1.0
  [1, 2]  =  1.0
  [2, 2]  =  -1.0
  [2, 3]  =  1.0
  [3, 3]  =  -1.0
  [3, 4]  =  1.0
  [4, 4]  =  0.0

julia> DifferentialOperator(x̄, (Reflecting(), Reflecting()), CentralSecondDifference())
4×4 SparseArrays.SparseMatrixCSC{Float64,Int64} with 10 stored entries:
  [1, 1]  =  -1.0
  [2, 1]  =  1.0
  [1, 2]  =  1.0
  [2, 2]  =  -2.0
  [3, 2]  =  1.0
  [2, 3]  =  1.0
  [3, 3]  =  -2.0
  [4, 3]  =  1.0
  [3, 4]  =  1.0
  [4, 4]  =  -1.0

source

SimpleDifferentialOperators.ExtensionDifferentialOperator — Method.

ExtensionDifferentialOperator(x̄, method::DifferenceMethod)

Returns a discretized differential operator of length(x̄) by length(x̄) + 2 matrix whose first and last columns are applied to the ghost nodes just before x̄[1] and x̄[end] respectively under no boundary condition using finite difference method specified by method.

Examples

julia> x̄ = 0:5
0:5

julia> ExtensionDifferentialOperator(x̄, BackwardFirstDifference())
4×6 SparseArrays.SparseMatrixCSC{Float64,Int64} with 8 stored entries:
  [1, 1]  =  -1.0
  [1, 2]  =  1.0
  [2, 2]  =  -1.0
  [2, 3]  =  1.0
  [3, 3]  =  -1.0
  [3, 4]  =  1.0
  [4, 4]  =  -1.0
  [4, 5]  =  1.0

julia> ExtensionDifferentialOperator(x̄, ForwardFirstDifference())
4×6 SparseArrays.SparseMatrixCSC{Float64,Int64} with 8 stored entries:
  [1, 2]  =  -1.0
  [1, 3]  =  1.0
  [2, 3]  =  -1.0
  [2, 4]  =  1.0
  [3, 4]  =  -1.0
  [3, 5]  =  1.0
  [4, 5]  =  -1.0
  [4, 6]  =  1.0

julia> ExtensionDifferentialOperator(x̄, CentralSecondDifference())
4×6 SparseArrays.SparseMatrixCSC{Float64,Int64} with 12 stored entries:
  [1, 1]  =  1.0
  [1, 2]  =  -2.0
  [2, 2]  =  1.0
  [1, 3]  =  1.0
  [2, 3]  =  -2.0
  [3, 3]  =  1.0
  [2, 4]  =  1.0
  [3, 4]  =  -2.0
  [4, 4]  =  1.0
  [3, 5]  =  1.0
  [4, 5]  =  -2.0
  [4, 6]  =  1.0

source

SimpleDifferentialOperators.L₁₊ — Method.

L₁₊(x̄)

Returns a discretized first-order differential operator of length(x̄) by length(x̄) + 2 matrix using forward difference under no boundary condition.

The first and last columns are applied to the ghost nodes just before x̄[1] and x̄[end] respectively.

Examples

julia> x̄ = 0:5
0:5

julia> Array(L₁₊(x̄))
4×6 Array{Float64,2}:
 0.0  -1.0   1.0   0.0   0.0  0.0
 0.0   0.0  -1.0   1.0   0.0  0.0
 0.0   0.0   0.0  -1.0   1.0  0.0
 0.0   0.0   0.0   0.0  -1.0  1.0

source

SimpleDifferentialOperators.L₁₊bc — Method.

L₁₊bc(x̄, bc::Tuple{BoundaryCondition, BoundaryCondition})

Returns a discretized first-order differential operator of length(x̄) by length(x̄) matrix using forward difference under boundary conditions specified by bc.

The first element of bc is applied to the lower bound, and second element of bc to the upper.

Examples

julia> x̄ = 0:5
0:5

julia> L₁₊bc(x̄, (Reflecting(), Reflecting()))
4×4 SparseArrays.SparseMatrixCSC{Float64,Int64} with 7 stored entries:
  [1, 1]  =  -1.0
  [1, 2]  =  1.0
  [2, 2]  =  -1.0
  [2, 3]  =  1.0
  [3, 3]  =  -1.0
  [3, 4]  =  1.0
  [4, 4]  =  0.0

source

SimpleDifferentialOperators.L₁₋ — Method.

L₁₋(x̄)

Returns a discretized first-order differential operator of length(x̄) by length(x̄) + 2 matrix using backward difference under no boundary condition.

The first and last columns are applied to the ghost nodes just before x̄[1] and x̄[end] respectively.

Examples

julia> x̄ = 1:3
1:3

julia> Array(L₁₋(x̄))
1×3 Array{Float64,2}:
 -1.0  1.0  0.0

source

SimpleDifferentialOperators.L₁₋bc — Method.

L₁₋bc(x̄, bc::Tuple{BoundaryCondition, BoundaryCondition})

Returns a discretized first-order differential operator of length(x̄) by length(x̄) matrix using backward difference under boundary conditions specified by bc.

The first element of bc is applied to the lower bound, and second element of bc to the upper.

Examples

julia> x̄ = 0:5
0:5

julia> L₁₋bc(x̄, (Reflecting(), Reflecting()))
4×4 SparseArrays.SparseMatrixCSC{Float64,Int64} with 7 stored entries:
  [1, 1]  =  0.0
  [2, 1]  =  -1.0
  [2, 2]  =  1.0
  [3, 2]  =  -1.0
  [3, 3]  =  1.0
  [4, 3]  =  -1.0
  [4, 4]  =  1.0

source

SimpleDifferentialOperators.L₂ — Method.

L₂(x̄)

Returns a discretized second-order differential operator of length(x̄) by length(x̄) + 2 matrix using central difference under no boundary condition.

The first and last columns are applied to the ghost nodes just before x̄[1] and x̄[end] respectively.

Examples

julia> x̄ = 0:5
0:5 

julia> Array(L₂(x̄))
4×6 Array{Float64,2}:
 1.0  -2.0   1.0   0.0   0.0  0.0
 0.0   1.0  -2.0   1.0   0.0  0.0
 0.0   0.0   1.0  -2.0   1.0  0.0
 0.0   0.0   0.0   1.0  -2.0  1.0

source

SimpleDifferentialOperators.L₂bc — Method.

L₂bc(x̄, bc::Tuple{BoundaryCondition, BoundaryCondition})

Returns a discretized second-order differential operator of length(x̄) by length(x̄) matrix using central difference under boundary conditions specified by bc.

The first element of bc is applied to the lower bound, and second element of bc to the upper.

Examples

julia> x̄ = 0:5
0:5

julia> L₂bc(x̄, (Reflecting(), Reflecting()))
4×4 SparseArrays.SparseMatrixCSC{Float64,Int64} with 10 stored entries:
  [1, 1]  =  -1.0
  [2, 1]  =  1.0
  [1, 2]  =  1.0
  [2, 2]  =  -2.0
  [3, 2]  =  1.0
  [2, 3]  =  1.0
  [3, 3]  =  -2.0
  [4, 3]  =  1.0
  [3, 4]  =  1.0
  [4, 4]  =  -1.0

source

SimpleDifferentialOperators.extrapolatetoboundary — Method.

extrapolatetoboundary(v, x̄, bc::Tuple{Mixed, Mixed})

Returns a length(x̄)-vector whose 2:(length(x̄)-1) elements are v, the first and last element are extrapolated v on the boundaries of x̄ according to boundary conditions bc given.

The first element of bc is applied to the lower bound, and second element of bc to the upper.

Examples

julia> x̄ = 0:5
0:5

julia> x = interiornodes(x̄)
1:4

julia> v = (x -> x^2).(x)
4-element Array{Int64,1}:
  1
  4
  9
 16

julia> extrapolatetoboundary(v, x̄, (Mixed(ξ = 1), Mixed(ξ = 1)))
6-element Array{Float64,1}:
 Inf
   1.0
   4.0
   9.0
  16.0
   8.0

source

SimpleDifferentialOperators.extrapolatetoboundary — Method.

extrapolatetoboundary(v, x̄, bc::Tuple{Reflecting, Reflecting})

Returns a length(x̄)-vector whose 2:(length(x̄)-1) elements are v, the first and last element are extrapolated v on the boundaries of x̄ according to boundary conditions bc given.

The first element of bc is applied to the lower bound, and second element of bc to the upper.

Examples

julia> x̄ = 0:5
0:5

julia> x = interiornodes(x̄)
1:4

julia> v = (x -> x^2).(x)
4-element Array{Int64,1}:
  1
  4
  9
 16

julia> extrapolatetoboundary(v, x̄, (Reflecting(), Reflecting()))
6-element Array{Int64,1}:
  1
  1
  4
  9
 16
 16

source

SimpleDifferentialOperators.interiornodes — Method.

interiornodes(x̄, bc)

Returns an interior grid corresponding to the boundary condition bc given extended grid x̄.

julia> x̄ = 0:5
0:5

julia> interiornodes(x̄, (Reflecting(), Reflecting()))
1:4

julia> x̄ = [1.0; 1.5; 1.7]
3-element Array{Float64,1}:
 1.0
 1.5
 1.7

julia> interiornodes(x̄, (Mixed(ξ = 1.0), Mixed(ξ = 1.0)))
1-element Array{Float64,1}:
 1.5

source

SimpleDifferentialOperators.interiornodes — Method.

interiornodes(x̄)

Returns an interior grid of length length(x̄)-2 given extended grid x̄.

julia> x̄ = 0:5
0:5

julia> interiornodes(x̄)
1:4

julia> x̄ = [1.0; 1.5; 1.7]
3-element Array{Float64,1}:
 1.0
 1.5
 1.7

julia> interiornodes(x̄)
1-element Array{Float64,1}:
 1.5

source