API

API

SimpleDifferentialOperators.DifferentialOperator

SimpleDifferentialOperators.DifferentialOperator

SimpleDifferentialOperators.ExtensionDifferentialOperator

SimpleDifferentialOperators.L̄₁₊

SimpleDifferentialOperators.L̄₁₋

SimpleDifferentialOperators.L̄₂

SimpleDifferentialOperators.L₁₊

SimpleDifferentialOperators.L₁₋

SimpleDifferentialOperators.L₂

SimpleDifferentialOperators.x̄

SimpleDifferentialOperators.DifferentialOperator — Method.

DifferentialOperator(x, bc::Tuple{Mixed, Mixed}, method::DifferenceMethod)

Returns a discretized differential operator of length(x) by length(x) matrix under mixed boundary conditions from bc using finite difference method specified by method.

Examples

julia> x = 1:3
1:3

julia> DifferentialOperator(x, (Mixed(1.0), Mixed(1.0)), BackwardFirstDifference())
3×3 LinearAlgebra.Tridiagonal{Float64,Array{Float64,1}}:
 -1.0   0.0   ⋅
 -1.0   1.0  0.0
   ⋅   -1.0  1.0

julia> DifferentialOperator(x, (Mixed(1.0), Mixed(1.0)), ForwardFirstDifference())
3×3 LinearAlgebra.Tridiagonal{Float64,Array{Float64,1}}:
 -1.0   1.0    ⋅
  0.0  -1.0   1.0
   ⋅    0.0  -1.0

julia> DifferentialOperator(x, (Mixed(1.0), Mixed(1.0)), CentralSecondDifference())
3×3 LinearAlgebra.Tridiagonal{Float64,Array{Float64,1}}:
 0.0   1.0    ⋅
 1.0  -2.0   1.0
  ⋅    1.0  -2.0

source

SimpleDifferentialOperators.DifferentialOperator — Method.

DifferentialOperator(x, bc::Tuple{Reflecting, Reflecting}, method::DifferenceMethod)

Returns a discretized differential operator of length(x) by length(x) matrix under reflecting boundary conditions from bc using finite difference method specified by method.

Examples

julia> x = 1:3
1:3

julia> DifferentialOperator(x, (Reflecting(), Reflecting()), BackwardFirstDifference())
3×3 LinearAlgebra.Tridiagonal{Float64,Array{Float64,1}}:
  0.0   0.0   ⋅
 -1.0   1.0  0.0
   ⋅   -1.0  1.0

julia> DifferentialOperator(x, (Reflecting(), Reflecting()), ForwardFirstDifference())
3×3 LinearAlgebra.Tridiagonal{Float64,Array{Float64,1}}:
 -1.0   1.0   ⋅
  0.0  -1.0  1.0
   ⋅    0.0  0.0

julia> DifferentialOperator(x, (Reflecting(), Reflecting()), CentralSecondDifference())
3×3 LinearAlgebra.Tridiagonal{Float64,Array{Float64,1}}:
 -1.0   1.0    ⋅
  1.0  -2.0   1.0
   ⋅    1.0  -1.0

source

SimpleDifferentialOperators.ExtensionDifferentialOperator — Method.

ExtensionDifferentialOperator(x, method::DifferenceMethod)

Returns a discretized differential operator of length(x) by length(x) + 2 matrix whose first and last columns are applied to the ghost nodes just before x[1] and x[end] respectively under no boundary condition using finite difference method specified by method.

Examples

julia> x = 1:3
1:3

julia> ExtensionDifferentialOperator(x, BackwardFirstDifference())
3×5 SparseArrays.SparseMatrixCSC{Float64,Int64} with 8 stored entries:
  [1, 1]  =  -1.0
  [1, 2]  =  1.0
  [2, 2]  =  -1.0
  [1, 3]  =  0.0
  [2, 3]  =  1.0
  [3, 3]  =  -1.0
  [2, 4]  =  0.0
  [3, 4]  =  1.0

julia> ExtensionDifferentialOperator(x, ForwardFirstDifference())
3×5 SparseArrays.SparseMatrixCSC{Float64,Int64} with 8 stored entries:
  [1, 2]  =  -1.0
  [2, 2]  =  0.0
  [1, 3]  =  1.0
  [2, 3]  =  -1.0
  [3, 3]  =  0.0
  [2, 4]  =  1.0
  [3, 4]  =  -1.0
  [3, 5]  =  1.0

julia> ExtensionDifferentialOperator(x, CentralSecondDifference())
3×5 SparseArrays.SparseMatrixCSC{Float64,Int64} with 9 stored entries:
  [1, 1]  =  1.0
  [1, 2]  =  -2.0
  [2, 2]  =  1.0
  [1, 3]  =  1.0
  [2, 3]  =  -2.0
  [3, 3]  =  1.0
  [2, 4]  =  1.0
  [3, 4]  =  -2.0
  [3, 5]  =  1.0

source

SimpleDifferentialOperators.L̄₁₊ — Method.

L̄₁₊(x)

Returns a discretized first-order differential operator of length(x) by length(x) + 2 matrix using forward difference under no boundary condition.

The first and last columns are applied to the ghost nodes just before x[1] and x[end] respectively.

Examples

julia> x = 1:3
1:3

julia> Array(L̄₁₊(x))
3×5 Array{Float64,2}:
 0.0  -1.0   1.0   0.0  0.0
 0.0   0.0  -1.0   1.0  0.0
 0.0   0.0   0.0  -1.0  1.0

source

SimpleDifferentialOperators.L̄₁₋ — Method.

L̄₁₋(x)

Returns a discretized first-order differential operator of length(x) by length(x) + 2 matrix using backward difference under no boundary condition.

The first and last columns are applied to the ghost nodes just before x[1] and x[end] respectively.

Examples

julia> x = 1:3
1:3

julia> Array(L̄₁₋(x))
3×5 Array{Float64,2}:
 -1.0   1.0   0.0  0.0  0.0
  0.0  -1.0   1.0  0.0  0.0
  0.0   0.0  -1.0  1.0  0.0

source

SimpleDifferentialOperators.L̄₂ — Method.

L̄₂(x)

Returns a discretized second-order differential operator of length(x) by length(x) + 2 matrix using central difference under no boundary condition.

The first and last columns are applied to the ghost nodes just before x[1] and x[end] respectively.

Examples

julia> x = 1:3
1:3

julia> Array(L̄₂(x))
3×5 Array{Float64,2}:
 1.0  -2.0   1.0   0.0  0.0
 0.0   1.0  -2.0   1.0  0.0
 0.0   0.0   1.0  -2.0  1.0

source

SimpleDifferentialOperators.L₁₊ — Method.

L₁₊(x, bc::Tuple{BoundaryCondition, BoundaryCondition})

Returns a discretized first-order differential operator of length(x) by length(x) matrix using forward difference under boundary conditions specified by bc.

The first element of bc is applied to the lower bound, and second element of bc to the upper.

Examples

julia> x = 1:3
1:3

julia> L₁₊(x, (Reflecting(), Reflecting()))
3×3 LinearAlgebra.Tridiagonal{Float64,Array{Float64,1}}:
 -1.0   1.0   ⋅
  0.0  -1.0  1.0
   ⋅    0.0  0.0

source

SimpleDifferentialOperators.L₁₋ — Method.

L₁₋(x, bc::Tuple{BoundaryCondition, BoundaryCondition})

Returns a discretized first-order differential operator of length(x) by length(x) matrix using backward difference under boundary conditions specified by bc.

The first element of bc is applied to the lower bound, and second element of bc to the upper.

Examples

julia> x = 1:3
1:3

julia> L₁₋(x, (Reflecting(), Reflecting()))
3×3 LinearAlgebra.Tridiagonal{Float64,Array{Float64,1}}:
  0.0   0.0   ⋅
 -1.0   1.0  0.0
   ⋅   -1.0  1.0

source

SimpleDifferentialOperators.L₂ — Method.

L₂(x, bc::Tuple{BoundaryCondition, BoundaryCondition})

Returns a discretized second-order differential operator of length(x) by length(x) matrix using central difference under boundary conditions specified by bc.

The first element of bc is applied to the lower bound, and second element of bc to the upper.

Examples

julia> x = 1:3
1:3

julia> L₂(x, (Reflecting(), Reflecting()))
3×3 LinearAlgebra.Tridiagonal{Float64,Array{Float64,1}}:
 -1.0   1.0    ⋅
  1.0  -2.0   1.0
   ⋅    1.0  -1.0

source

SimpleDifferentialOperators.x̄ — Method.

x̄(x)

Returns an extended grid of length length(x)+2 given grid x.

The first and last elements of the returned extended grid represent the ghost nodes just before x[1] and x[end] respectively.

julia> x = 1:3
1:3

julia> x̄(x)
5-element Array{Int64,1}:
 0
 1
 2
 3
 4

julia> x = [1.0; 1.5; 1.7]
3-element Array{Float64,1}:
 1.0
 1.5
 1.7

julia> x̄(x)
5-element Array{Float64,1}:
 0.5
 1.0
 1.5
 1.7
 1.9

source